WLAN configuration (static address): reject IPv6 addresses.
Without this, entering an IPv6 address as the interface address (or possibly
also others, such as gateway address) will, on connection to that WLAN,
trigger an uncaught exception:

*** FATAL EXCEPTION IN SYSTEM PROCESS: WifiStateMachine
java.lang.IllegalArgumentException: command '59 interface setcfg wlan0 2001:db8::ffff 24 up' failed with '501 59 Flag unsupported'
	at com.android.server.NetworkManagementService.setInterfaceConfig(NetworkManagementService.java:546)
	at android.net.wifi.WifiStateMachine$ObtainingIpState.enter(WifiStateMachine.java:3449)
	at com.android.internal.util.StateMachine$SmHandler.invokeEnterMethods(StateMachine.java:958)
[etc.]

followed by a restart of the UI then another connection attempt. So long as
the misconfigured-for AP is present (and preferred?), this will continue to
happen.

Change-Id: I524480339c37ad079fd21d90a6b1fc5864089cb5